Fully replace serialization.ts with the typed version
